pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eqn2sim"
version = "0.1.0"
description = "Deterministic equation-to-Simulink compiler and MATLAB bridge"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Chancel Lavoie" }
]
dependencies = [
  "Flask>=3.0",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "openai>=1.0",
  "pydantic>=2.0",
  "python-dotenv>=1.0",
  "scipy>=1.12",
  "sympy>=1.12",
]

[project.optional-dependencies]
dev = [
  "hypothesis>=6.0",
  "pre-commit>=4.0",
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.12.0",
]

[project.scripts]
eqn2sim-bootstrap = "scripts.bootstrap_distribution:main"
eqn2sim-gui = "scripts.run_eqn2sim_gui:main"
eqn2sim-lint = "scripts.run_lint:main"
eqn2sim-pipeline = "pipeline.run_pipeline:main"
eqn2sim-tests = "scripts.run_tests:main"
eqn2sim-examples = "scripts.run_examples:main"
eqn2sim-benchmark = "scripts.generate_benchmark:main"
eqn2sim-frontdoor-benchmark = "scripts.generate_frontdoor_benchmark:main"
eqn2sim-full-benchmark = "scripts.run_full_benchmark:main"
eqn2sim-layout-benchmark = "scripts.generate_layout_benchmark:main"

[tool.setuptools]
py-modules = ["repo_paths"]

[tool.setuptools.packages.find]
include = [
  "backend",
  "backend.*",
  "canonicalize",
  "canonicalize.*",
  "eqn2sim_gui",
  "eqn2sim_gui.*",
  "ir",
  "ir.*",
  "latex_frontend",
  "latex_frontend.*",
  "pipeline",
  "pipeline.*",
  "scripts",
  "scripts.*",
  "simulate",
  "simulate.*",
  "simucompilebench",
  "simucompilebench.*",
  "simulink",
  "simulink.*",
  "states",
  "states.*",
]
exclude = [
  "*.tests",
  "*.tests.*",
  "tests",
  "tests.*",
]

[tool.setuptools.package-data]
eqn2sim_gui = ["templates/*.html"]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = [
  "workspace",
]

[tool.ruff.lint]
select = [
  "E7",
  "E9",
  "F",
]
